+++ tests/unit/test_conditions.py
@@ -989,7 +989,7 @@ def test_ior_basic():
     # in
     (InCondition(column=c, values=[]), "(M.c in [])"),
     (InCondition(column=c, values=[2, 3]), "(M.c in [2, 3])"),
-    (InCondition(column=c, values=[MockColumn("d"), 3]), "(M.c in [<MockColumn[M.d]>, 3])"),
+    (InCondition(column=c, values=[MockColumn("name"), 3]), "(M.c in [<MockColumn[M.name:d_name]>, 3])"),
 
     # empty
     (Condition(), "()")